Filename changed to DigestCalculator
